+++ app/src/main/java/com/github/mikelambert/killswitch/ui/devices/DevicesFragment.java
@@ -1,4 +1,4 @@
-package com.github.mikelambert.killswitch.ui.home;
+package com.github.mikelambert.killswitch.ui.devices;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,17 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.github.mikelambert.killswitch.R;
 
-public class HomeFragment extends Fragment {
+public class DevicesFragment extends Fragment {
 
-    private HomeViewModel homeViewModel;
+    private DevicesViewModel devicesViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        homeViewModel =
-                ViewModelProviders.of(this).get(HomeViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_home, container, false);
-        final TextView textView = root.findViewById(R.id.text_home);
-        homeViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        devicesViewModel =
+                ViewModelProviders.of(this).get(DevicesViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_devices, container, false);
+        final TextView textView = root.findViewById(R.id.text_dashboard);
+        devicesViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
