+++ app/src/main/java/com/github/mikelambert/killswitch/ui/devices/DevicesViewModel.java
@@ -1,16 +1,16 @@
-package com.github.mikelambert.killswitch.ui.home;
+package com.github.mikelambert.killswitch.ui.devices;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class DevicesViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public DevicesViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is devices fragment");
     }
 
     public LiveData<String> getText() {
