+++ app/src/main/java/com/github/mikelambert/killswitch/ui/status/StatusViewModel.java
@@ -1,16 +1,16 @@
-package com.github.mikelambert.killswitch.ui.notifications;
+package com.github.mikelambert.killswitch.ui.status;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class StatusViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public StatusViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is status fragment");
     }
 
     public LiveData<String> getText() {
